# artifacts/api-server/src/routes/magicLinks.ts
/**
 * Magic Link API
 *
 * POST /magic-links/generate — creates a signed token (auth-protected)
 * GET  /magic-links/redeem/:token — validates and redeems the token (public)
 *
 * TTLs:
 *  login          15 minutes
 *  payment        24 hours
 *  appointment    24 hours
 *  accountability 48 hours
 */
import { Router, Request, Response } from "express";
import { db } from "@workspace/db";
import { magicLinks, planProgressLogs, profiles, usersTable } from "@workspace/db";
import { eq, and, gt, isNull } from "drizzle-orm";
import { z } from "zod";
import { randomUUID } from "crypto";
import { sendEmail, sendNotification } from "../lib/comms";
import { magicLinkEmail } from "../emails/magic-link";
import { paymentConfirmedEmail } from "../emails/payment-confirmed";
import { createSession, SESSION_COOKIE } from "../lib/auth";

const MAGIC_LINK_SESSION_TTL = 15 * 60 * 1000; // 15 minutes for magic-link login

const BASE_URL = process.env.BASE_URL || "https://healthplanfactory.com";

const router = Router();

const ACTION_TTLS_MS: Record<string, number> = {
  login: 15 * 60 * 1000,
  payment: 24 * 60 * 60 * 1000,
  appointment: 24 * 60 * 60 * 1000,
  accountability: 48 * 60 * 60 * 1000,
};

const GenerateMagicLinkBody = z.object({
  action: z.enum(["login", "payment", "appointment", "accountability"]),
  payload: z.record(z.unknown()).optional().default({}),
  sendEmail: z.boolean().optional().default(false),
});

/**
 * POST /api/magic-links/generate
 * Auth-protected. Creates a signed token for the given action.
 */
router.post("/magic-links/generate", async (req: Request, res: Response) => {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Authentication required" });
    return;
  }

  const body = GenerateMagicLinkBody.safeParse(req.body);
  if (!body.success) {
    res.status(400).json({ error: "Validation error", details: body.error.flatten() });
    return;
  }

  const requestingUser = req.user!;
  const isAdmin = requestingUser.role === "admin";
  const { action, payload, sendEmail: shouldEmail } = body.data;

  // payment and appointment actions can only be generated by admins (server-side).
  // Members generating these themselves would bypass payment/appointment controls.
  if ((action === "payment" || action === "appointment") && !isAdmin) {
    res.status(403).json({ error: "Only admins can generate payment or appointment magic links" });
    return;
  }

  // For member-generated links (login, accountability), profileId is the requesting user.
  // For admin-generated links, allow targeting any profileId via payload.targetProfileId.
  const profileId =
    isAdmin && typeof payload?.targetProfileId === "string"
      ? payload.targetProfileId
      : requestingUser.id;

  const ttl = ACTION_TTLS_MS[action] ?? ACTION_TTLS_MS.appointment;
  const expiresAt = new Date(Date.now() + ttl);

  try {
    const [link] = await db
      .insert(magicLinks)
      .values({
        profileId,
        action,
        payload,
        expiresAt,
      })
      .returning();

    const url = `${BASE_URL}/api/magic-links/redeem/${link.id}`;

    if (shouldEmail) {
      const [profile] = await db
        .select({ email: profiles.email, displayName: profiles.displayName })
        .from(profiles)
        .where(eq(profiles.id, profileId))
        .limit(1);

      if (profile?.email) {
        const expiresInMinutes = Math.round(ttl / 60000);
        const { subject, html } = magicLinkEmail({
          displayName: profile.displayName,
          action,
          magicLinkUrl: url,
          expiresInMinutes,
        });
        sendEmail(profileId, profile.email, subject, html, "magic-link").catch(() => {});
      }
    }

    res.status(201).json({ token: link.id, url, expiresAt: link.expiresAt });
  } catch (err) {
    const message = err instanceof Error ? err.message : "Internal server error";
    res.status(500).json({ error: message });
  }
});

/**
 * GET /api/magic-links/redeem/:token
 * Public. Validates the token, marks it used, and redirects to the appropriate deep-link.
 */
router.get("/magic-links/redeem/:token", async (req: Request, res: Response) => {
  const { token } = req.params;

  if (!token || typeof token !== "string") {
    res.status(400).json({ error: "Invalid token" });
    return;
  }

  try {
    const now = new Date();

    // Atomic single-use enforcement: UPDATE ... WHERE id=? AND used_at IS NULL AND expires_at > now RETURNING *
    // If two concurrent requests race, only one gets a returned row; the other gets empty and returns 410.
    const [link] = await db
      .update(magicLinks)
      .set({ usedAt: now })
      .where(
        and(
          eq(magicLinks.id, token),
          gt(magicLinks.expiresAt, now),
          isNull(magicLinks.usedAt),
        ),
      )
      .returning();

    if (!link) {
      res.status(410).json({ error: "Magic link is invalid, expired, or already used" });
      return;
    }

    const payload = (link.payload ?? {}) as Record<string, unknown>;

    // All magic-link actions create a short-lived authenticated session so the user
    // lands in the correct screen without needing to log in separately.
    const [profile] = await db
      .select({ id: profiles.id, email: profiles.email, role: profiles.role, avatarUrl: profiles.avatarUrl })
      .from(profiles)
      .where(eq(profiles.id, link.profileId))
      .limit(1);

    const [userRow] = profile
      ? await db
        .select({ firstName: usersTable.firstName, lastName: usersTable.lastName })
        .from(usersTable)
        .where(eq(usersTable.id, link.profileId))
        .limit(1)
      : [undefined];

    if (profile) {
      const sessionData = {
        user: {
          id: profile.id,
          email: profile.email ?? null,
          firstName: userRow?.firstName ?? null,
          lastName: userRow?.lastName ?? null,
          profileImageUrl: profile.avatarUrl ?? null,
          role: profile.role ?? "member",
        },
        access_token: "",
      };
      const sid = await createSession(sessionData, MAGIC_LINK_SESSION_TTL);
      res.cookie(SESSION_COOKIE, sid, {
        httpOnly: true,
        secure: true,
        sameSite: "lax",
        path: "/",
        maxAge: MAGIC_LINK_SESSION_TTL,
      });
    }

    if (link.action === "payment") {
      // Action: Payment confirmation — mark subscription as active and send payment-confirmed email
      try {
        const amountFormatted = typeof payload.amountFormatted === "string" ? payload.amountFormatted : "$9.99/month";
        const description = typeof payload.description === "string" ? payload.description : "Monthly subscription";

        await db.update(profiles).set({ subscriptionStatus: "plus", updatedAt: new Date() }).where(eq(profiles.id, link.profileId));

        const [profile] = await db
          .select({ email: profiles.email, displayName: profiles.displayName })
          .from(profiles)
          .where(eq(profiles.id, link.profileId))
          .limit(1);

        if (profile?.email) {
          const { subject, html } = paymentConfirmedEmail({
            displayName: profile.displayName,
            amountFormatted,
            description,
            dashboardUrl: `${BASE_URL}/dashboard`,
          });
          sendNotification({
            profileId: link.profileId,
            email: profile.email,
            type: "payment-confirmed",
            subject,
            html,
            smsBody: `Health Plan Factory: Payment of ${amountFormatted} confirmed. Thank you!`,
          }).catch(() => {});
        }
      } catch (paymentErr) {
        console.error("[magic-link] payment action side-effect error:", paymentErr);
      }
    } else if (link.action === "appointment") {
      // Action: Appointment acceptance — record the confirmed appointment as a session log
      // and update the magic link payload to mark it as confirmed.
      try {
        await db.insert(planProgressLogs).values({
          id: randomUUID(),
          profileId: link.profileId,
          planId: typeof payload.planId === "string" ? payload.planId : null,
          modalityId: typeof payload.modalityId === "string" ? payload.modalityId : null,
          note: "Appointment confirmed via magic link",
          sessionDate: typeof payload.appointmentDate === "string"
            ? new Date(payload.appointmentDate)
            : now,
        });
      } catch (apptErr) {
        console.error("[magic-link] appointment confirmation error:", apptErr);
      }
    } else if (link.action === "accountability") {
      // Action: Accountability check-in — log a minimal check-in progress entry
      try {
        await db.insert(planProgressLogs).values({
          id: randomUUID(),
          profileId: link.profileId,
          planId: typeof payload.planId === "string" ? payload.planId : null,
          note: "Accountability check-in via magic link",
          sessionDate: now,
        });
      } catch (checkinErr) {
        console.error("[magic-link] accountability check-in error:", checkinErr);
      }
    }

    const destination = getRedirectDestination(link.action, payload);
    res.redirect(302, destination);
  } catch (err) {
    const message = err instanceof Error ? err.message : "Internal server error";
    res.status(500).json({ error: message });
  }
});

function sanitizeReturnTo(returnTo: unknown): string {
  if (typeof returnTo !== "string") return "/dashboard";
  const trimmed = returnTo.trim();
  if (!trimmed.startsWith("/") || trimmed.startsWith("//")) return "/dashboard";
  try {
    const parsed = new URL(trimmed, "https://placeholder.internal");
    if (parsed.origin !== "https://placeholder.internal") return "/dashboard";
    return parsed.pathname + parsed.search + parsed.hash;
  } catch {
    return "/dashboard";
  }
}

function getRedirectDestination(
  action: string,
  payload: Record<string, unknown>,
): string {
  switch (action) {
    case "login":
      return sanitizeReturnTo(payload.returnTo);
    case "payment":
      return "/dashboard?action=payment-confirmed";
    case "appointment":
      return "/dashboard?action=appointment-accepted";
    case "accountability":
      return "/progress?action=checkin";
    default:
      return "/dashboard";
  }
}

export default router;